+++ spec/lib/waterdrop/instrumentation/vendors/datadog/metrics_listener_spec.rb
@@ -265,10 +265,15 @@ RSpec.describe_current do
     before do
       producer.produce_sync(topic: rand.to_s, payload: rand.to_s)
       sleep(1)
+      listener.send(:report_metric, metric, statistics)
     end
 
-    let(:metric) { described_class::RdKafkaMetric.new(:gauge, :topics, 'topics.batchcnt.avg', %w[batchcnt avg]) }
     let(:guages) { dummy_client.buffer[:gauge] }
+
+    let(:metric) do
+      described_class::RdKafkaMetric.new(:gauge, :topics, 'topics.batchcnt.avg', %w[batchcnt avg])
+    end
+
     let(:statistics) do
       {
         'name' => 'producer-1',
@@ -283,8 +288,6 @@ RSpec.describe_current do
         }
       }
     end
-
-    before { listener.send(:report_metric, metric, statistics) }
 
     it 'report metric will publish statistic with topic tag' do
       expect(guages['waterdrop.topics.batchcnt.avg']).to eql [[6956, { tags: ['topic:test'] }]]