+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 2.7.3 (Unreleased)
 - [Enhancement] Provide `#idempotent?` similar to `#transactional?`.
 - [Enhancement] Provide alias to `#with` named `#variant`.
+- [Fix] Prevent from creating `acks` altering variants on idempotent producers.
 
 ## 2.7.2 (2024-05-09)
 - [Fix] Fix missing requirement of `delegate` for non-Rails use-cases. Always require delegate for variants usage (samsm)